+++ CHANGELOG.md
@@ -24,6 +24,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Consider using https://github.com/ascopes/protobuf-maven-plugin (more up to date) 
 - Multi-release-jar?
 
+### Changed
+- Cleaned up test topologies.  [#145](https://github.com/scionproto-contrib/jpan/pull/145)
+
 ## [0.4.1] - 2024-11-22
 
 ### Fixed